chore: Improve project structure for Aviator modules
