+++ fcli-core/fcli-aviator-common/src/main/java/com/fortify/cli/aviator/audit/IssueObjBuilder.java
@@ -1,13 +1,13 @@
-package com.fortify.cli.aviator.core;
+package com.fortify.cli.aviator.audit;
 
 import java.util.HashSet;
 import java.util.Set;
 
-import com.fortify.cli.aviator.core.model.AnalysisInfo;
-import com.fortify.cli.aviator.core.model.File;
-import com.fortify.cli.aviator.core.model.IssueData;
-import com.fortify.cli.aviator.core.model.UserPrompt;
-import com.fortify.cli.aviator.fpr.Vulnerability;
+import com.fortify.cli.aviator.audit.model.AnalysisInfo;
+import com.fortify.cli.aviator.audit.model.File;
+import com.fortify.cli.aviator.audit.model.IssueData;
+import com.fortify.cli.aviator.audit.model.UserPrompt;
+import com.fortify.cli.aviator.fpr.model.Vulnerability;
 import com.fortify.cli.aviator.util.FileTypeLanguageMapperUtil;
 import com.fortify.cli.aviator.util.FileUtil;
 import com.fortify.cli.aviator.util.StringUtil;
