+++ fcli-core/fcli-aviator-common/src/main/java/com/fortify/cli/aviator/audit/model/AuditResult.java
@@ -1,4 +1,4 @@
-package com.fortify.cli.aviator.core.model;
+package com.fortify.cli.aviator.audit.model;
 
 
 import javax.xml.bind.annotation.XmlElement;
