+++ fcli-core/fcli-aviator-common/src/main/java/com/fortify/cli/aviator/config/ExtensionsConfig.java
@@ -1,10 +1,11 @@
-package com.fortify.cli.aviator.util;
+package com.fortify.cli.aviator.config;
 
 import java.util.HashMap;
 import java.util.Map;
 
 import com.formkiq.graalvm.annotations.Reflectable;
 
+import com.fortify.cli.aviator.util.StringUtil;
 import lombok.NoArgsConstructor;
 
 @NoArgsConstructor @Reflectable
