+++ fcli-core/fcli-aviator-common/src/main/java/com/fortify/cli/aviator/fpr/model/Vulnerability.java
@@ -1,10 +1,10 @@
-package com.fortify.cli.aviator.fpr;
+package com.fortify.cli.aviator.fpr.model;
 
 import java.util.List;
 
 import com.formkiq.graalvm.annotations.Reflectable;
-import com.fortify.cli.aviator.core.model.File;
-import com.fortify.cli.aviator.core.model.StackTraceElement;
+import com.fortify.cli.aviator.audit.model.File;
+import com.fortify.cli.aviator.audit.model.StackTraceElement;
 
 import lombok.Builder;
 import lombok.Getter;
